Move `download_file` helpers to `tests/support/pytest/helpers.py`
